Card and CardCarousel components added
